requirements.txt
joblib==1.2.0
matplotlib==3.6.3
numpy==1.23.5
opencv_python_headless==4.5.4.60
pandas==1.5.3
Pillow==9.5.0
scikit_learn==1.2.2
streamlit==1.22.0